End of section 3

+++ AuctionService/DTOs/AuctionDTO.cs
@@ -3,11 +3,11 @@
     public class AuctionDTO
     {
         public Guid Id { get; set; }
-        public int ReservePrice { get; set; }
+        public Int64 ReservePrice { get; set; }
         public string Seller { get; set; }
         public string Winner { get; set; }
-        public int SoldAmount { get; set; }
-        public int CurrentHighBid { get; set; }
+        public Int64 SoldAmount { get; set; }
+        public Int64 CurrentHighBid { get; set; }
         public DateTime CreatedDate { get; set; }
         public DateTime UpdatedDate { get; set; }
         public DateTime AuctionEnd { get; set; }
